:recycle: client type 분리
